// src/order-management/enums/Text.tsx
export enum FilterText {
    SEARCH_PALACE_HOLDER = '검색어를 입력해 주세요',
    ORDER_NUM = '가자고 주문번호',
    OPT_ORDER_NUM = '가자고 옵션 주문번호',
    ITEM_ORDER_NUM = '가자고 아이템 주문번호',
    PARTNER_ORDER_NUM = '구매처 주문번호',
    PURCHASER_NAME = '구매자명',
    PURCHASER_PHONE_NUM = '구매자 전화번호',
    USER_NAME = '이용자명',
    USER_PHONE_NUM = '이용자 전화번호',
    DEAL_ID = '딜 ID',
    DEAL_NAME = '딜 명',
    INIT = '초기화',
    SEARCH = '조회',
    SEARCH_CONDITION = '검색조건',
    SEARCH_PERIOD = '조회기간',
    PARTNER_NAME = '구매처명',
    USE_STATUS = '이용상태',
    REFUND_STATUS = '환불상태',
    THIRD_PARTY_NAME = '연동사명',
    PURCHASE_DATE_BASE = '구매일 기준',
    USE_DATE_BASE = '사용일 기준',
    REFUND_REQUEST_DATE_BASE = '환불 요청일 기준',
    START_DATE = '시작일',
    END_DATE = '종료일',
    TODAY = '오늘',
    ONE_MONTH = '1개월',
    THREE_MONTH = '3개월',
    SIX_MONTH = '6개월',
    QUANTITY = '수량',
    TOTAL_PRICE = '총 주문금액',
    PURCHASE_DATE = '구매일시',
    NOTICE = '검색 조건을 입력하세요.',
    NO_RESULT = '검색 결과가 없습니다.',
    DATE_FORMAT = 'YYYY.MM.DD'
}